vr_project/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
Pillow>=9.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
hnswlib>=0.7.0
streamlit>=1.28.0
ultralytics>=8.0.0
open_clip_torch>=2.20.0
accelerate>=0.24.0
ftfy>=6.1.1
regex>=2023.0.0
tqdm>=4.65.0
pandas>=2.0.0
matplotlib>=3.7.0
Requests>=2.31.0
scipy>=1.11.0
